Add entry header selecting C++11 or C++17
