+++ bench/RESULTS.md
@@ -28,3 +28,38 @@ BM_Boost256Multiplication          519 ns          519 ns        26950
 The looped benchmark makes timing differences more visible. `wide_integer` has
 slightly faster additions than the compiler builtin or Boost types, while its
 256‑bit multiplication remains faster than Boost's implementation.
+
+## C++11 vs C++17
+
+Benchmarking the same operations under different C++ standards shows comparable performance. Sample results from the `performance.cpp` benchmark:
+
+### C++17
+```
+BM_Addition             1.62 ns         1.62 ns      7709981
+BM_Subtraction          2.06 ns         2.06 ns      6982822
+BM_Multiplication       8.09 ns         8.09 ns      1729611
+BM_Division             99.6 ns         99.6 ns       137063
+BM_ToStringOld        179251 ns       179247 ns           76
+BM_ToStringNew         97955 ns        97953 ns          135
+```
+
+### C++11
+```
+BM_Addition             1.78 ns         1.78 ns      7415415
+BM_Subtraction          2.03 ns         2.03 ns      6982707
+BM_Multiplication       8.12 ns         8.12 ns      1722390
+BM_Division             97.7 ns         97.7 ns       147232
+BM_ToStringOld        185689 ns       185693 ns           76
+BM_ToStringNew         98485 ns        98481 ns          140
+```
+
+| Benchmark      | C++17 (ns) | C++11 (ns) | C++11/17 |
+|---------------|-----------:|-----------:|---------:|
+| Addition       |       1.62 |       1.78 |    1.10x |
+| Subtraction    |       2.06 |       2.03 |    0.99x |
+| Multiplication |       8.09 |       8.12 |    1.00x |
+| Division       |       99.6 |       97.7 |    0.98x |
+| ToStringOld    |    179251  |    185689  |    1.04x |
+| ToStringNew    |     97955  |     98485  |    1.01x |
+
+Arithmetic operations are within the same order of magnitude, with timing differences generally under 10%.
