+++ bench/performance.cpp
@@ -1,10 +1,5 @@
 #include <benchmark/benchmark.h>
-
-#ifdef USE_CXX11_HEADER
-#    include <wide_integer/wide_integer_cxx11.h>
-#else
-#    include <wide_integer/wide_integer.h>
-#endif
+#include <wide_integer/wide_integer.h>
 
 using WInt = wide::integer<256, unsigned>;
 
@@ -14,8 +9,11 @@ static void BM_Addition(benchmark::State & state)
     WInt b = 987654321;
     for (auto _ : state)
     {
-        WInt c = a + b;
+        benchmark::DoNotOptimize(a);
+        benchmark::DoNotOptimize(b);
+        auto c = a + b;
         benchmark::DoNotOptimize(c);
+        benchmark::ClobberMemory();
     }
 }
 
@@ -25,8 +23,11 @@ static void BM_Subtraction(benchmark::State & state)
     WInt b = 123456789;
     for (auto _ : state)
     {
-        WInt c = a - b;
+        benchmark::DoNotOptimize(a);
+        benchmark::DoNotOptimize(b);
+        auto c = a - b;
         benchmark::DoNotOptimize(c);
+        benchmark::ClobberMemory();
     }
 }
 
@@ -36,8 +37,11 @@ static void BM_Multiplication(benchmark::State & state)
     WInt b = 987654321;
     for (auto _ : state)
     {
-        WInt c = a * b;
+        benchmark::DoNotOptimize(a);
+        benchmark::DoNotOptimize(b);
+        auto c = a * b;
         benchmark::DoNotOptimize(c);
+        benchmark::ClobberMemory();
     }
 }
 
@@ -47,8 +51,11 @@ static void BM_Division(benchmark::State & state)
     WInt b = 123456;
     for (auto _ : state)
     {
-        WInt c = a / b;
+        benchmark::DoNotOptimize(a);
+        benchmark::DoNotOptimize(b);
+        auto c = a / b;
         benchmark::DoNotOptimize(c);
+        benchmark::ClobberMemory();
     }
 }
 
@@ -77,8 +84,10 @@ static void BM_ToStringOld(benchmark::State & state)
     WInt a = (WInt(1) << 255) + 123456789;
     for (auto _ : state)
     {
+        benchmark::DoNotOptimize(a);
         auto s = to_string_slow(a);
         benchmark::DoNotOptimize(s);
+        benchmark::ClobberMemory();
     }
 }
 
@@ -87,8 +96,10 @@ static void BM_ToStringNew(benchmark::State & state)
     WInt a = (WInt(1) << 255) + 123456789;
     for (auto _ : state)
     {
+        benchmark::DoNotOptimize(a);
         auto s = wide::to_string(a);
         benchmark::DoNotOptimize(s);
+        benchmark::ClobberMemory();
     }
 }
